+++ opentelemetry-instrumentation/tests/test_instrumentor.py
@@ -54,11 +54,12 @@ class TestInstrumentor(TestCase):
     def test_singleton(self):
         self.assertIs(self.Instrumentor(), self.Instrumentor())
 
+    @patch("opentelemetry.instrumentation.instrumentor._LOG")
     @patch(
         "opentelemetry.instrumentation.instrumentor.BaseInstrumentor._check_dependency_conflicts"
     )
     def test_instrument_missing_dependency_raise(
-        self, mock__check_dependency_conflicts
+        self, mock__check_dependency_conflicts, mock_logger
     ):
         instrumentor = self.Instrumentor()
 
@@ -71,3 +72,21 @@ class TestInstrumentor(TestCase):
             instrumentor.instrument,
             raise_exception_on_conflict=True,
         )
+        mock_logger.error.assert_not_called()
+
+    @patch("opentelemetry.instrumentation.instrumentor._LOG")
+    @patch(
+        "opentelemetry.instrumentation.instrumentor.BaseInstrumentor._check_dependency_conflicts"
+    )
+    def test_instrument_missing_dependency_log_error(
+        self, mock__check_dependency_conflicts, mock_logger
+    ):
+        instrumentor = self.Instrumentor()
+        conflict = DependencyConflict(
+            "missing", "missing"
+        )
+        mock__check_dependency_conflicts.return_value = conflict
+        self.assertIsNone(
+            instrumentor.instrument(raise_exception_on_conflict=False)
+        )
+        mock_logger.error.assert_any_call(conflict)